+++ tests/router_core_tests.py
@@ -416,7 +416,7 @@ class RouterTests(TestCase):
             return msg.x + msg.y, msg.x, msg.y
 
         self.assertEqual(
-            {'sum': 3, 'x': 1, 'y': 2},
+            ({'sum': 3, 'x': 1, 'y': 2},),
             add(x=1, y=2)
         )
 
@@ -434,6 +434,20 @@ class RouterTests(TestCase):
                 {'combination': 'stuffiest'}
             ),
             suffixes(pre='stuff', sufs=['y', 'ier', 'iest'])
+        )
+
+    def test_live_mode(self):
+        'adding emit_immediately emits immediately'
+        @self.router.node(['x'], emit_immediately=True)
+        def node(msg):
+            yield 1
+            yield 2
+            yield 3
+
+        # return value is identical
+        self.assertEqual(
+            ({'x': 1}, {'x': 2}, {'x': 3},),
+            node(TODO=1),
         )
 
     def test_routing(self):