+++ docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+0.5.0
+-----
+
+ - Nodes which return single values will now be wrapped in a tuple, for
+   consistency with generator nodes. Routing will proceed as before.
+ - You can now pass "emit_immediately" into a node definition to immediately
+   route your message, without waiting for the rest of the generator to finish.
+
 0.4.0
 -----